+++ src/main/java/blackjack/domain/Players.java
@@ -5,6 +5,7 @@ import blackjack.domain.participant.Player;
 import java.util.List;
 import java.util.Objects;
 import java.util.function.Consumer;
+import java.util.function.Supplier;
 
 public class Players {
     private final List<Player> players;
@@ -25,9 +26,23 @@ public class Players {
         return new Players(players);
     }
 
-   public void forEach(Consumer<Player> action){
+    public void deal(Supplier<TrumpCard> cardSupplier) {
+        forEach(player -> {
+            player.hit(cardSupplier.get());
+            player.hit(cardSupplier.get());
+        });
+    }
+
+    public void betPlayers(BetDecision decision){
+        players.forEach(player -> {
+            String amount = decision.decideBet(player.getName());
+            player.placeBet(amount);
+        });
+    }
+
+    public void forEach(Consumer<Player> action){
         players.forEach(action);
-   }
+    }
 
     public int count() {
         return players.size();
